requirements.txt
fastapi==0.120.2
pydantic==2.12.3
uvicorn==0.38.0

google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1

passlib==1.7.4
email-validator
typing-extensions>=4.14.1
